+++ src/components/SpruceForm/types.ts
@@ -1,4 +1,4 @@
-import { FieldValidation } from "@rjsf/core";
+import { FieldValidation, FormProps, Field } from "@rjsf/core";
 
 // typescript utility to recursively iterate through an object and add a method called addError to each property
 type RecursivelyAddError<T> = T extends object
@@ -12,3 +12,19 @@ export type ValidateProps<T> = (
   FormState: T,
   errors: RecursivelyAddError<T>
 ) => RecursivelyAddError<T>;
+
+type CustomFormatFields = {
+  jiraHost?: string;
+};
+
+export type SpruceFormProps<A = any> = Pick<
+  FormProps<A>,
+  "schema" | "onChange" | "formData"
+> &
+  Partial<FormProps<A>> & { customFormatFields?: CustomFormatFields };
+
+export type GetFormSchema = (...any) => {
+  fields: Record<string, Field>;
+  schema: SpruceFormProps["schema"];
+  uiSchema: SpruceFormProps["uiSchema"];
+};
