+++ src/pages/task/taskTabs/buildBaronAndAnnotations/AddIssueModal/index.tsx
@@ -2,8 +2,9 @@ import { useState } from "react";
 import { useMutation } from "@apollo/client";
 import { useAnnotationAnalytics } from "analytics";
 import { ConfirmationModal } from "components/ConfirmationModal";
-import { SpruceForm, SpruceFormProps } from "components/SpruceForm";
+import { SpruceForm } from "components/SpruceForm";
 import { AccordionFieldTemplate } from "components/SpruceForm/FieldTemplates";
+import { SpruceFormProps } from "components/SpruceForm/types";
 import { useToastContext } from "context/toast";
 import {
   AddAnnotationIssueMutation,
